+++ misc/dbt-materialize/dbt/include/materialize/macros/utils/listagg.sql
@@ -13,20 +13,12 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 
-{% macro mz_generate_name(identifier) -%}
-    {{ exceptions.warn(
+{% macro materialize__listagg(measure, delimiter_text, order_by_clause, limit_num) -%}
+    {{ exceptions.raise_compiler_error(
         """
-        The mz_generate_name macro is deprecated and will be removed in a future release of dbt-materialize.
-        Please use the {{ this }} jinja function to reference the relation instead:
+        Materialize supports the list_agg() function; use it directly.
 
-        e.g.
-        {{ config(materialized=''source'') }}
-
-        CREATE SOURCE {{ this }} ...
-
-        Will create a source using the filename as the identifier.
-        Documentation for {{ this }} can be found: https://docs.getdbt.com/reference/dbt-jinja-functions/this
+        For details, see: https://materialize.com/docs/sql/functions/list_agg/
         """
     )}}
-  {{ return("{}.{}.{}".format(database, schema, identifier)) }}
-{% endmacro %}
+{%- endmacro %}